Add map-select component for displaying maps
